backend/requirements.txt
asgiref==3.4.1
Django==3.2.19
django-cors==0.1
django-cors-headers==3.10.1
django-rest==0.8.7
django-restframework==0.0.1
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.1.1
